refactor: migrate user record helpers to three-tier test architecture
Move insertUserRow, deleteUserRow, insertTestUser, seedUserCacheEntry,
insertUserCacheEntry, insertVm0ApiKeys to db-test-seeders/users.ts with
@why-db-direct JSDoc.

Move getUserRow, countUserRows to db-test-assertions/users.ts.

Delete 2 unused functions: getTestUserPreferences, setTestUserPreferences.

Re-exports in api-test-helpers/users.ts preserve backward compatibility.

Closes #9344
